Fix search init with aggs from url
